feat: introduce auto slice video
fix #233
