+++ .dockerignore
@@ -6,6 +6,7 @@ pnpm-lock.yaml
 **/__pycache__
 test/*
 Videos/*
+!Videos/.gitkeep
 settings-production.toml
 startRecord-production.sh
 src/utils/cookies.json
